Refactor Memperbaiki penamaan file dan variabel
