+++ lib/constants.dart
@@ -1,4 +1,5 @@
 // Jika menggunakan android device (AVD)
 // String url = 'http://10.0.0.2:8000/api/';
-String url = 'http://127.0.0.1:8000/api/';
+String baseUrl = 'http://127.0.0.1:8000/';
+String apiUrl = '{$baseUrl}api/';
 var headers = {'Accept': 'application/json'};
