+++ lib/views/dashboard.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:get/get.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/authentication.dart';
-import 'package:parkquest_parkir_gamifikasi/Controllers/leaderboard_controller.dart';
+import 'package:parkquest_parkir_gamifikasi/Controllers/AuthenticationController.dart';
+import 'package:parkquest_parkir_gamifikasi/Controllers/LeaderboardController.dart';
 import 'package:parkquest_parkir_gamifikasi/views/widgets/leaderboard_top_three.dart';
 
 class Dashboard extends StatelessWidget {
